requirements.txt
sentence-transformers
bm25
pinecone
spacy
rank-bm25
python-dotenv
psycopg2
pymupdf
langchain_groq
fastapi 
"uvicorn[standard]"
pydantic
python-multipart
langchain_mistralai
pandas
sqlalchemy
asyncpg